feat: publish external work only after durability
